Refactor feature specs into smaller units
